# Fix the implementation of the Bilinear layer