The current implementation causes an unnecessary huge copy of `x1` in the first matmul (it copies it `output_dims` times). This fixes it by changing the order of the dims in the weight so it can be done without any copy at all.

The outputs are identical provided one transposes the weight from one to the other. Even for modest sizes of `128` it gives a 20x speedup on my M2 air, measured using the following:

```python
import mlx.nn as nn
import mlx.core as mx
def bench(f, *args):
    ys = []
    for i in range(10):
        ys.append(f(*args))
    mx.eval(ys)

m = nn.Bilinear(128, 128, 128)
x = mx.random.normal((64, 128))
y = mx.random.normal((64, 128))
get_ipython().run_line_magic('timeit', 'bench(m, x, y)')
```

Before
```
49.8 ms ± 493 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

Now
```
2.52 ms ± 6.54 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```